chore: solve circular dependencies
